refactor: move setEmptyFieldsToUndefined to separate file
